# api-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.announcementItem:
    properties:
      content:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      is_active:
        type: boolean
      title:
        type: string
    type: object
...
  handlers.createAnnouncementRequest:
    properties:
      content:
        type: string
      expires_at:
        type: string
      image_url:
        type: string
      title:
        type: string
    type: object
...
  handlers.createNodeGroupRequest:
    properties:
      name:
        type: string
    type: object
...
  handlers.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      totp_code:
        type: string
    type: object
  handlers.loginResponse:
    properties:
      token:
        type: string
    type: object
...
  handlers.patchAlertRequest:
    properties:
      ack:
        type: boolean
      close:
        type: boolean
      silence_minutes:
        type: integer
    type: object
...
  handlers.terminateSessionRequest:
    properties:
      cooldown_minutes:
        type: integer
    type: object
...
  services.Entry:
    properties:
      actor_id:
        type: string
      actor_label:
        type: string
      actor_type:
        type: string
      created_at:
        type: string
      detail:
        additionalProperties: {}
        type: object
      event_type:
        type: string
      id:
        type: string
      severity:
        type: string
      target_id:
        type: string
      target_type:
        type: string
    type: object
...
host: localhost:2053
info:
  contact: {}
  description: VPN Panel Management API
  title: Proxima VPN API
  version: 0.1.0
paths:
  /admin/activity:
    get:
      description: Returns the most recent recorded panel events, newest first. Supplying
        both target_type and target_id narrows the feed to that target.
      parameters:
      - description: Maximum entries to return (default 20)
        in: query
        name: limit
        type: integer
      - description: Restrict to one target type, e.g. node (requires target_id)
        in: query
        name: target_type
        type: string
      - description: Restrict to one target id (requires target_type)
        in: query
        name: target_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.Entry'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get recent activity
      tags:
      - admin-stats
  /admin/alerts/{id}:
    patch:
      consumes:
      - application/json
      description: Acknowledging removes an alert from the notifiable count while
        leaving it visible. Silencing suppresses notifications for a bounded window
        without stopping evaluation. Closing resolves an alert the evaluator cannot
        resolve itself, and is accepted only for a stale alert or one whose node no
        longer exists.
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      - description: Acknowledge flag, silence duration in minutes, or close flag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.patchAlertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Acknowledge, silence or close an alert
      tags:
      - admin
  /admin/announcements:
    get:
      description: Returns all announcements for admin management
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.announcementItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List announcements
      tags:
      - admin-announcements
    post:
      consumes:
      - application/json
      description: Creates a new announcement shown in the user portal
      parameters:
      - description: Announcement to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createAnnouncementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.announcementItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create announcement
      tags:
      - admin-announcements
...
  /admin/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate admin with email/password and optional TOTP
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.loginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Admin login
      tags:
      - admin-auth
  /admin/backup/download:
    get:
      description: Downloads a backup file by key (defaults to latest)
      parameters:
      - description: Backup key
        in: query
        name: key
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download backup
      tags:
      - admin-backup
  /admin/backup/list:
    get:
      description: Returns a list of available backups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List backups
      tags:
      - admin-backup
  /admin/backup/trigger:
    post:
      description: Triggers a manual database backup
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Trigger backup
      tags:
      - admin-backup
  /admin/devices/{id}/terminate:
    post:
      consumes:
      - application/json
      description: Evicts a device so the node agent withdraws its credential on the
        next config poll
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Cooldown in minutes
        in: body
        name: body
        schema:
          $ref: '#/definitions/handlers.terminateSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Terminate a device's session
      tags:
      - admin-stats
  /admin/inbounds/{id}:
    delete:
      description: Removes an inbound by ID
      parameters:
      - description: Inbound ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete inbound
      tags:
      - admin-inbounds
    put:
      consumes:
      - application/json
      description: Partially updates an inbound configuration
      parameters:
      - description: Inbound ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateInboundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.inboundResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update inbound
      tags:
      - admin-inbounds
  /admin/inbounds/{id}/toggle:
    put:
      description: Toggles the enabled/disabled state of an inbound
      parameters:
      - description: Inbound ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.inboundResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Toggle inbound
      tags:
      - admin-inbounds
  /admin/node-groups:
    get:
      description: Returns all node groups with their node counts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.nodeGroupListItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List node groups
      tags:
      - admin-node-groups
    post:
      consumes:
      - application/json
      description: Creates a new node group
      parameters:
      - description: Node group name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createNodeGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.nodeGroupListItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create node group
      tags:
      - admin-node-groups
...
  /admin/nodes:
    get:
      description: Returns all nodes including pending (awaiting registration) ones
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.nodeListItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List nodes
      tags:
      - admin-nodes
  /admin/nodes/{id}:
    delete:
      description: Removes a node by ID
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete node
      tags:
      - admin-nodes
    get:
      description: Returns a single node by ID
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.nodeListItem'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get node
      tags:
      - admin-nodes
    put:
      consumes:
      - application/json
      description: Partially updates a node's name, country, or region. Cannot edit
        pending nodes.
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateNodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update node
      tags:
      - admin-nodes
...
  /admin/nodes/{nodeId}/inbounds:
    get:
      description: Returns all inbounds for a given node
      parameters:
      - description: Node ID
        in: path
        name: nodeId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.inboundResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List inbounds
      tags:
      - admin-inbounds
    post:
      consumes:
      - application/json
      description: Adds a new inbound configuration for a node
      parameters:
      - description: Node ID
        in: path
        name: nodeId
        required: true
        type: string
      - description: Inbound configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createInboundRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.inboundResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create inbound
      tags:
      - admin-inbounds
...
  /admin/online-users:
    get:
      description: Returns detailed info about currently connected users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties:
                type: string
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get online users
      tags:
      - admin-stats
  /admin/orders:
    get:
      description: Returns all plan orders, optionally filtered by status
      parameters:
      - description: Filter by status (pending, paid, cancelled)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.adminOrderItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List plan orders
      tags:
      - admin-orders
...
  /admin/orders/{id}/cancel:
    post:
      description: Admin cancels a pending order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel a plan order
      tags:
      - admin-orders
...
  /admin/plans:
    get:
      description: Returns all plans with node group information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.planResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List plans
      tags:
      - admin-plans
    post:
      consumes:
      - application/json
      description: Creates a new subscription plan
      parameters:
      - description: Plan details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createPlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.planResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create plan
      tags:
      - admin-plans
...
  /admin/stats:
    get:
      description: Returns aggregate statistics for the admin dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get dashboard stats
      tags:
      - admin-stats
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"